requirements.txt
colorama==0.4.4
texttable
biopython